Enhance code quality and validations

+++ src/Controllers/CompanyManagementController.java
@@ -34,6 +34,10 @@ public class CompanyManagementController {
         return v;
     };
     public Function<Void, Void> assignEmployeeToDepartmentManager = v -> {
+        if (!hasAnyPersonnels(service.getPersonnels())) {
+            return v;
+        }
+
         List<Employee> unmanagedEmployees = service.getListOf(Employee.class, e -> e.getManagerId() == null);
         service.displayTableOfPersonnels("LIST OF UNMANAGED EMPLOYEES", unmanagedEmployees);
         String employeeId = userInteractor.readLine("Enter employee's ID you want to assign to be managed: ");
@@ -59,23 +63,28 @@ public class CompanyManagementController {
         return v;
     };
     public Function<Void, Void> addNewPersonnel = v -> {
-        List<Personnel> personnels = service.getPersonnels();
-        personnels = company.addNewPersonel(personnels);
-        service.savePersonnels(personnels);
+        company.addNewPersonel();
         return v;
     };
     public Function<Void, Void> deletePersonnel = v -> {
+        if (!hasAnyPersonnels(service.getPersonnels())) {
+            return v;
+        }
+
         String employeeId = userInteractor
                 .readLine("Enter ID of Employee you want to delete: ");
         Employee employee = (Employee) service.findPersonnel(p -> p.getId().equals(employeeId));
+        if (employee == null) {
+            userInteractor.displayMessage("Process stop. Cannot find a personnel with ID " + employeeId);
+            return v;
+        }
         employee.delete();
         return v;
     };
     public Function<Void, Void> displayAllPersonnel = v -> {
         List<Personnel> personnels = service.getPersonnels();
-        if (personnels.size() == 0) {
-            userInteractor
-                    .displayMessage("Personnels: Empty list.\nThere is no personnel in the company info currently.\n");
+
+        if (!hasAnyPersonnels(personnels)) {
             return v;
         }
 
@@ -85,6 +94,10 @@ public class CompanyManagementController {
 
     public Function<Void, Void> calculateTotalSalaries = v -> {
         List<Personnel> personnels = service.getPersonnels();
+        if (!hasAnyPersonnels(personnels)) {
+            return v;
+        }
+
         BigDecimal total = BigDecimal.ZERO;
         for (Personnel personnel : personnels) {
             total = total.add(personnel.calculateMonthlySalary());
@@ -96,10 +109,16 @@ public class CompanyManagementController {
 
     public Function<Void, Void> findHighestSalaryEmployee = v -> {
         List<Personnel> personnels = service.getPersonnels();
+
+        if (!hasAnyPersonnels(personnels)) {
+            return v;
+        }
+
         BigDecimal maxSalary = BigDecimal.ZERO;
         Personnel p = personnels.get(0);
         for (Personnel personnel : personnels) {
-            if (maxSalary.compareTo(personnel.calculateMonthlySalary()) < 0) {
+            if (personnel instanceof Employee &&
+                    maxSalary.compareTo(personnel.calculateMonthlySalary()) < 0) {
                 p = personnel;
             }
         }
@@ -111,11 +130,20 @@ public class CompanyManagementController {
 
     public Function<Void, Void> findManagerWithMostManagedEmployees = v -> {
         List<Personnel> personnels = service.getPersonnels();
+
+        if (!hasAnyPersonnels(personnels)) {
+            return v;
+        }
+
         Personnel dm = personnels.stream().filter(p -> p.getIsDeptManager()).findFirst().get();
+        if (dm == null) {
+            userInteractor.displayMessage("--> Currently, there is no Department Manager in the company personnels.\n");
+            return v;
+        }
         DepartmentManager maxDm = (DepartmentManager) dm;
         DepartmentManager temp;
         for (Personnel personnel : personnels) {
-            if (personnel.getIsDeptManager()) {
+            if (personnel instanceof DepartmentManager) {
                 temp = (DepartmentManager) personnel;
                 if (maxDm.getTotalManagedEmployees() < temp.getTotalManagedEmployees()) {
                     maxDm = temp;
@@ -131,6 +159,11 @@ public class CompanyManagementController {
 
     public Function<Void, Void> sortPersonnelNamesAlphabetically = v -> {
         List<Personnel> personnels = service.getPersonnels();
+
+        if (!hasAnyPersonnels(personnels)) {
+            return v;
+        }
+
         personnels.sort(new SortByName());
 
         service.displayTableOfPersonnels("Personnels sorted by name alphabetically:", personnels);
@@ -139,6 +172,12 @@ public class CompanyManagementController {
 
     public Function<Void, Void> sortPersonnelsSalaryDesc = v -> {
         List<Personnel> personnels = service.getPersonnels();
+
+        if (!hasAnyPersonnels(personnels)) {
+            return v;
+        }
+        ;
+
         personnels.sort(new SortByMonthSalaryDesc());
 
         service.displayTableOfPersonnels("Personnels sorted by salary descendingly:", personnels);
@@ -147,12 +186,22 @@ public class CompanyManagementController {
 
     public Function<Void, Void> findLargestSharePercentageDirector = v -> {
         List<Personnel> personnels = service.getPersonnels();
-        Director maxDirector = (Director) personnels.stream().filter(p -> p.getIsDirector()).findFirst().get();
+
+        if (!hasAnyPersonnels(personnels)) {
+            return v;
+        }
+
+        Personnel personnel = personnels.stream().filter(p -> p.getIsDirector()).findFirst().get();
+        if (personnel == null) {
+            userInteractor.displayMessage("--> Currently, there is no Director in the company personnels.\n");
+            return v;
+        }
+        Director maxDirector = (Director) personnel;
         Director temp;
 
-        for (Personnel personnel : personnels) {
-            if (personnel.getIsDirector()) {
-                temp = (Director) personnel;
+        for (Personnel p : personnels) {
+            if (p instanceof Director) {
+                temp = (Director) p;
                 if (maxDirector.getSharePercentage() < temp.getSharePercentage()) {
                     maxDirector = temp;
                 }
@@ -167,16 +216,27 @@ public class CompanyManagementController {
 
     public Function<Void, Void> calculateEveryDirectorIncome = v -> {
         List<Personnel> personnels = service.getPersonnels();
-        userInteractor.displayMessage("Monthly income of every Director:\n");
+        if (!hasAnyPersonnels(personnels)) {
+            return v;
+        }
         Director temp;
+        int counter = 0;
+
+        userInteractor.displayMessage("Calculating monthly income of every director...\n");
         for (Personnel personnel : personnels) {
             if (personnel.getIsDirector()) {
                 temp = (Director) personnel;
 
                 userInteractor.displayMessage(temp.getFullName() + ": "
                         + temp.calculateMonthlyIncome(company).toPlainString());
+                ++counter;
             }
         }
+
+        if (counter == 0) {
+            userInteractor.displayMessage("There is no director in the company personnels.\n");
+        }
+
         return v;
     };
 
@@ -184,4 +244,12 @@ public class CompanyManagementController {
         app.stop();
         return v;
     };
+
+    private boolean hasAnyPersonnels(List<Personnel> personnels) {
+        if (personnels == null || personnels.size() == 0) {
+            userInteractor.displayMessage("Process stopped. Currently, there is no personnel in the company data.");
+            return false;
+        }
+        return true;
+    }
 }
